Move CayenneCatalogProvider to cayenne crate
