+++ crates/cayenne/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 authors = ["Spice.ai OSS Authors"]
 description = "Cayenne: A DuckLake-inspired lakehouse format using SQLite for metadata and Vortex files for data storage"
-edition = "2021"
+edition = "2024"
 license = "Apache-2.0"
 name = "cayenne"
 repository = "https://github.com/spiceai/spiceai"
@@ -34,6 +34,7 @@ object_store = { workspace = true }
 parking_lot = { workspace = true }
 roaring = { workspace = true }
 runtime-datafusion = { path = "../runtime-datafusion" }
+runtime-table-partition = { path = "../runtime-table-partition", optional = true }
 rusqlite = { workspace = true }
 serde = { workspace = true }
 tokio-rusqlite = { workspace = true }
@@ -60,6 +61,7 @@ aegis = { version = "0.9.3", default-features = false, features = [
 [features]
 default = []
 turso = ["dep:turso"]
+partition-table-provider = ["dep:runtime-table-partition"]
 
 [dev-dependencies]
 criterion = { version = "0.7", features = ["html_reports"] }
