+++ crates/cayenne/benches/metastore_operations.rs
@@ -6,7 +6,7 @@ use cayenne::metastore::turso::TursoMetastore;
 use cayenne::metastore::{
     ExecuteParams, MetastoreBackend, MetastoreValue, QueryParams, QueryRowParams,
 };
-use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
+use criterion::{BenchmarkId, Criterion, criterion_group, criterion_main};
 use std::hint::black_box;
 use tempfile::TempDir;
 use tokio::runtime::Runtime;
