+++ crates/cayenne/src/provider/delete.rs
@@ -44,12 +44,12 @@ mod sink;
 mod vector_io;
 
 // Public API - re-exported in provider/mod.rs
-pub(crate) use sink::file_based::FileBasedDeletionSink;
 pub use sink::CayenneDeletionSink;
+pub(crate) use sink::file_based::FileBasedDeletionSink;
 
 // Crate-internal types used by table.rs
 pub(crate) use filter_exec::{Int64PkDeletionFilterExec, KeyBasedDeletionFilterExec};
 pub(crate) use vector_io::{
-    detect_deletion_type_and_read, DeletionIdentifier, DeletionVectorWriteSpec,
-    DeletionVectorWriter,
+    DeletionIdentifier, DeletionVectorWriteSpec, DeletionVectorWriter,
+    detect_deletion_type_and_read,
 };
