+++ crates/cayenne/src/provider/delete/sink/position_based.rs
@@ -22,27 +22,27 @@ limitations under the License.
 
 use super::super::vector_io::{DeletionIdentifier, DeletionVectorWriteSpec, DeletionVectorWriter};
 use super::CayenneDeletionSink;
+use crate::provider::Error;
 use crate::provider::constants::DELETION_CACHE_LOCK_POISONED;
 use crate::provider::utils::convert_to_u64_box;
-use crate::provider::Error;
 use datafusion::datasource::listing::ListingTable;
 use datafusion::execution::context::SessionContext;
 use datafusion::optimizer::analyzer::type_coercion::TypeCoercionRewriter;
+use datafusion_common::DFSchema;
 use datafusion_common::tree_node::TreeNode;
 use datafusion_common::utils::get_available_parallelism;
-use datafusion_common::DFSchema;
-use datafusion_expr::execution_props::ExecutionProps;
 use datafusion_expr::Expr;
+use datafusion_expr::execution_props::ExecutionProps;
 use datafusion_physical_expr::create_physical_expr;
 use futures::StreamExt;
 use object_store::ObjectStore;
 use roaring::{RoaringBitmap, RoaringTreemap};
 use std::collections::HashMap;
 use std::sync::{Arc, LazyLock};
+use vortex::VortexSessionDefault;
 use vortex::array::arrow::IntoArrowArray;
 use vortex::file::OpenOptionsSessionExt;
 use vortex::layout::layouts::row_idx::row_idx;
-use vortex::VortexSessionDefault;
 use vortex_datafusion::DefaultExpressionConvertor;
 use vortex_datafusion::ExpressionConvertor;
 use vortex_session::VortexSession;
