+++ crates/cayenne/src/provider/mod.rs
@@ -61,7 +61,7 @@ pub use deletion_strategy::{PkDeletionStrategy, PkDeletionStrategyWithCache};
 pub use retention::TimeRetentionFilterBuilder;
 pub use scan::CayenneAccelerationExec;
 pub use table::{CayenneTableProvider, CayenneTableProviderBuilder};
-pub use vortex_format::{attach_deletion_vectors_to_config, DeletionFilteringVortexFormat};
+pub use vortex_format::{DeletionFilteringVortexFormat, attach_deletion_vectors_to_config};
 
 // Re-export deletion utilities for advanced use cases
 pub use delete::CayenneDeletionSink;
@@ -138,7 +138,9 @@ pub enum Error {
     #[snafu(display("Internal error in table '{table}': {message}"))]
     Internal { table: String, message: String },
 
-    #[snafu(display("Unable to open Cayenne acceleration file ({file_path}). Too many Cayenne acceleration files are open. Try increasing your system's maximum open file count, or increase the size of generated Cayenne files with the parameter \"cayenne_target_file_size_mb\". For more details, visit: https://spiceai.org/docs/components/data-accelerators/cayenne#params"))]
+    #[snafu(display(
+        "Unable to open Cayenne acceleration file ({file_path}). Too many Cayenne acceleration files are open. Try increasing your system's maximum open file count, or increase the size of generated Cayenne files with the parameter \"cayenne_target_file_size_mb\". For more details, visit: https://spiceai.org/docs/components/data-accelerators/cayenne#params"
+    ))]
     TooManyOpenFiles { file_path: String },
 
     /// A previous write was interrupted, leaving the table in a potentially
@@ -836,7 +838,7 @@ mod tests {
             schema: Arc::clone(&schema),
             primary_key: vec!["email".to_string()],
             on_conflict: Some(OnConflict::Upsert(ColumnReference::new(vec![
-                "email".to_string()
+                "email".to_string(),
             ]))),
             base_path: data_dir.to_string_lossy().to_string(),
             partition_column: None,
@@ -981,7 +983,7 @@ mod tests {
             schema: Arc::clone(&schema),
             primary_key: vec!["email".to_string()],
             on_conflict: Some(OnConflict::Upsert(ColumnReference::new(vec![
-                "email".to_string()
+                "email".to_string(),
             ]))),
             base_path: data_dir.to_string_lossy().to_string(),
             partition_column: None,
