+++ crates/cayenne/src/provider/vortex_format.rs
@@ -36,11 +36,11 @@ use datafusion::datasource::file_format::FileFormat;
 use datafusion_catalog::Session;
 use datafusion_common::Result as DFResult;
 use datafusion_common::Statistics;
+use datafusion_datasource::PartitionedFile;
+use datafusion_datasource::TableSchema;
 use datafusion_datasource::file_compression_type::FileCompressionType;
 use datafusion_datasource::file_groups::FileGroup;
 use datafusion_datasource::file_scan_config::FileScanConfig;
-use datafusion_datasource::PartitionedFile;
-use datafusion_datasource::TableSchema;
 use datafusion_physical_plan::{DisplayAs, DisplayFormatType, ExecutionPlan, PlanProperties};
 use object_store::{ObjectMeta, ObjectStore};
 use roaring::{RoaringBitmap, RoaringTreemap};
