+++ crates/cayenne/tests/acid_compliance_test.rs
@@ -39,8 +39,8 @@ use arrow::array::{Array, Int64Array, RecordBatch, StringArray};
 use arrow::datatypes::{DataType, Field, Schema};
 
 use cayenne::{
-    metadata::CreateTableOptions, CayenneTableProvider, CayenneTableProviderBuilder,
-    MetadataCatalog,
+    CayenneTableProvider, CayenneTableProviderBuilder, MetadataCatalog,
+    metadata::CreateTableOptions,
 };
 
 use common::TestFixture;
