+++ crates/cayenne/tests/multi_partition_test.rs
@@ -34,8 +34,8 @@ mod common;
 use arrow::array::{Int64Array, StringArray};
 use arrow::datatypes::{DataType, Field, Schema};
 use arrow::record_batch::RecordBatch;
-use cayenne::metadata::{CreateTableOptions, PartitionMetadata};
 use cayenne::CayenneTableProvider;
+use cayenne::metadata::{CreateTableOptions, PartitionMetadata};
 use datafusion::prelude::*;
 use std::sync::Arc;
