+++ crates/cayenne/tests/read_time_filtering_test.rs
@@ -25,7 +25,7 @@ use arrow::array::{Int64Array, RecordBatch, StringArray};
 use arrow::datatypes::{DataType, Field, Schema};
 
 use cayenne::{
-    metadata::CreateTableOptions, CayenneCatalog, CayenneTableProvider, MetadataCatalog,
+    CayenneCatalog, CayenneTableProvider, MetadataCatalog, metadata::CreateTableOptions,
 };
 
 use datafusion::datasource::TableProvider;
@@ -40,8 +40,8 @@ use tempfile::TempDir;
 
 type TestResult<T> = Result<T, Box<dyn std::error::Error + Send + Sync>>;
 
-async fn setup_test_table(
-) -> TestResult<(Arc<CayenneTableProvider>, SessionContext, TempDir, TempDir)> {
+async fn setup_test_table()
+-> TestResult<(Arc<CayenneTableProvider>, SessionContext, TempDir, TempDir)> {
     // Create temporary directories for data and metadata
     let data_dir = TempDir::new()?;
     let metadata_dir = TempDir::new()?;
