+++ crates/runtime/Cargo.toml
@@ -357,7 +357,7 @@ turso = [
 ]
 vortex = ["ballista-core/vortex", "ballista-executor/vortex"]
 [target.'cfg(not(windows))'.dependencies]
-cayenne = { path = "../cayenne", features = ["turso"] }
+cayenne = { path = "../cayenne", features = ["turso", "partition-table-provider"] }
 vortex.workspace = true
 vortex-datafusion.workspace = true
