+++ crates/runtime/src/catalogconnector/cayenne/provider.rs
@@ -14,270 +14,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-//! Cayenne catalog and schema provider implementations for `DataFusion`.
-//!
-//! Provides a dynamic-namespace catalog backed by a Cayenne [`MetadataCatalog`]
-//! (`SQLite`) and local file storage for data files.
-//!
-//! Tables are organized into namespaces (schemas) specified at DDL time.
-//! In the metadata catalog, table names are stored with a namespace prefix
-//! (`namespace/table_name`) so that namespace membership survives restarts.
-
-use std::any::Any;
-use std::collections::HashMap;
-use std::sync::{Arc, RwLock};
-
 use async_trait::async_trait;
-use cayenne::metadata::VortexConfig;
-use cayenne::{CayenneCatalog, CayenneTableProvider, CayenneTableProviderBuilder, MetadataCatalog};
-use datafusion::catalog::{CatalogProvider, SchemaProvider, TableProvider};
-use datafusion::error::Result as DFResult;
-use datafusion::execution::runtime_env::RuntimeEnv;
-use snafu::prelude::*;
-
-use crate::component::catalog::Catalog;
-use crate::parameters::Parameters;
-use crate::spice_data_base_path;
-use data_components::RefreshableCatalogProvider;
-use data_components::poly::PolyTableProvider;
-use runtime_table_partition::provider::PartitionTableProvider;
+use cayenne::CayenneCatalogProvider;
 
 use crate::catalogconnector::PartitionAwareCatalog;
-
-#[derive(Debug, Snafu)]
-pub enum Error {
-    #[snafu(display("Failed to initialize Cayenne catalog: {source}"))]
-    CatalogInit {
-        source: cayenne::catalog::CatalogError,
-    },
-
-    #[snafu(display("Invalid Cayenne catalog configuration: {message}"))]
-    InvalidConfiguration { message: String },
-
-    #[snafu(display("Failed to create Cayenne table provider: {source}"))]
-    TableProvider {
-        source: cayenne::catalog::CatalogError,
-    },
-}
-
-pub type Result<T, E = Error> = std::result::Result<T, E>;
-
-/// Default catalog name used when no catalog ID is specified.
-const DEFAULT_CATALOG_NAME: &str = "cayenne";
-
-/// `DataFusion` [`CatalogProvider`] backed by a Cayenne metadata catalog.
-///
-/// Schemas are created dynamically when tables are created via DDL.
-/// There are no automatic "default" or "public" schemas; tables are only
-/// accessible through the namespace they were explicitly created in
-/// (e.g., `my_catalog.my_namespace.my_table`).
-///
-/// Data is stored on local disk, metadata in local `SQLite`.
-pub struct CayenneCatalogProvider {
-    /// The underlying Cayenne metadata catalog (`SQLite`).
-    catalog: Arc<dyn MetadataCatalog>,
-    /// Vortex configuration for table providers.
-    vortex_config: VortexConfig,
-    /// Base path for table data on local disk.
-    data_base_path: String,
-    /// Local metadata directory.
-    metadata_dir: String,
-    /// Shared `RuntimeEnv` from the main Spice runtime for cache coherence.
-    runtime_env: Arc<RuntimeEnv>,
-    /// Schema providers keyed by namespace name, created dynamically via DDL.
-    schemas: RwLock<HashMap<String, Arc<dyn SchemaProvider>>>,
-}
-
-impl std::fmt::Debug for CayenneCatalogProvider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("CayenneCatalogProvider")
-            .field("data_base_path", &self.data_base_path)
-            .field("metadata_dir", &self.metadata_dir)
-            .finish_non_exhaustive()
-    }
-}
-
-impl CayenneCatalogProvider {
-    /// Create a new Cayenne catalog provider.
-    ///
-    /// Initializes the `SQLite` metadata catalog and local file storage.
-    /// The `catalog_id` field is ignored; Cayenne always uses a fixed default name.
-    pub async fn try_new(
-        params: Parameters,
-        _catalog_config: &Catalog,
-        runtime_env: Arc<RuntimeEnv>,
-    ) -> Result<Self> {
-        let catalog_name = DEFAULT_CATALOG_NAME;
-
-        // Resolve metadata directory
-        let metadata_dir = params
-            .get("cayenne_metadata_dir")
-            .expose()
-            .ok()
-            .map_or_else(
-                || format!("{}/cayenne_{catalog_name}/metadata", spice_data_base_path()),
-                String::from,
-            );
-
-        // Ensure metadata directory exists
-        std::fs::create_dir_all(&metadata_dir).map_err(|e| Error::InvalidConfiguration {
-            message: format!("Failed to create metadata directory '{metadata_dir}': {e}"),
-        })?;
-
-        // Initialize SQLite catalog
-        let connection_string = format!("sqlite://{metadata_dir}/cayenne.db");
-        let catalog = Arc::new(
-            CayenneCatalog::new(connection_string).map_err(|e| Error::CatalogInit { source: e })?,
-        ) as Arc<dyn MetadataCatalog>;
-
-        catalog.init().await.context(CatalogInitSnafu)?;
-
-        // Initialize local file storage
-        let data_dir = params.get("cayenne_data_dir").expose().ok().map_or_else(
-            || format!("{}/cayenne_{catalog_name}/data", spice_data_base_path()),
-            String::from,
-        );
-
-        std::fs::create_dir_all(&data_dir).map_err(|e| Error::InvalidConfiguration {
-            message: format!("Failed to create data directory '{data_dir}': {e}"),
-        })?;
-
-        // Ensure trailing slash for consistent path joining
-        let data_base_path = if data_dir.ends_with('/') {
-            data_dir
-        } else {
-            format!("{data_dir}/")
-        };
-
-        tracing::info!("Cayenne catalog using local file storage at '{data_base_path}'");
-
-        // Parse Vortex config from parameters
-        let vortex_config = Self::parse_vortex_config(&params);
-
-        let provider = Self {
-            catalog,
-            vortex_config,
-            data_base_path,
-            metadata_dir,
-            runtime_env,
-            schemas: RwLock::new(HashMap::new()),
-        };
-
-        Ok(provider)
-    }
-
-    /// Returns a reference to the underlying Cayenne metadata catalog.
-    #[must_use]
-    pub fn metadata_catalog(&self) -> &Arc<dyn MetadataCatalog> {
-        &self.catalog
-    }
-
-    /// Returns the base data path on local disk.
-    #[must_use]
-    pub fn data_base_path(&self) -> &str {
-        &self.data_base_path
-    }
-
-    /// Returns the Vortex configuration.
-    #[must_use]
-    pub fn vortex_config(&self) -> &VortexConfig {
-        &self.vortex_config
-    }
-
-    /// Returns the shared `RuntimeEnv`.
-    #[must_use]
-    pub fn runtime_env(&self) -> &Arc<RuntimeEnv> {
-        &self.runtime_env
-    }
-
-    /// Returns the metadata directory path.
-    #[must_use]
-    pub fn metadata_dir(&self) -> &str {
-        &self.metadata_dir
-    }
-
-    /// Returns the schema provider for a namespace if it exists.
-    #[must_use]
-    pub fn schema_provider(&self, name: &str) -> Option<Arc<dyn SchemaProvider>> {
-        self.schemas
-            .read()
-            .ok()
-            .and_then(|schemas| schemas.get(name).cloned())
-    }
-
-    /// Registers or replaces a schema provider for a namespace.
-    pub fn register_schema_provider(
-        &self,
-        name: &str,
-        schema: Arc<dyn SchemaProvider>,
-    ) -> DFResult<Option<Arc<dyn SchemaProvider>>> {
-        match self.schemas.write() {
-            Ok(mut schemas) => Ok(schemas.insert(name.to_string(), schema)),
-            Err(_) => Err(datafusion::error::DataFusionError::Internal(
-                "Failed to acquire write lock on Cayenne schemas".to_string(),
-            )),
-        }
-    }
-
-    /// Parse Vortex configuration from catalog parameters.
-    fn parse_vortex_config(params: &Parameters) -> VortexConfig {
-        let mut config = VortexConfig::default();
-
-        if let Some(v) = params.get("cayenne_footer_cache_mb").expose().ok()
-            && let Ok(val) = v.parse::<usize>()
-        {
-            config.footer_cache_mb = val;
-        }
-        if let Some(v) = params.get("cayenne_segment_cache_mb").expose().ok()
-            && let Ok(val) = v.parse::<usize>()
-        {
-            config.segment_cache_mb = val;
-        }
-        if let Some(v) = params.get("cayenne_target_file_size_mb").expose().ok()
-            && let Ok(val) = v.parse::<usize>()
-        {
-            config.target_vortex_file_size_mb = val;
-        }
-        if let Some(v) = params.get("cayenne_compression_strategy").expose().ok() {
-            match v.to_lowercase().as_str() {
-                "zstd" => {
-                    config.compression_strategy = cayenne::metadata::CompressionStrategy::Zstd;
-                }
-                "btrblocks" => {
-                    config.compression_strategy = cayenne::metadata::CompressionStrategy::Btrblocks;
-                }
-                _ => {}
-            }
-        }
-
-        config
-    }
-}
-
-impl CatalogProvider for CayenneCatalogProvider {
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn schema_names(&self) -> Vec<String> {
-        self.schemas
-            .read()
-            .map(|schemas| schemas.keys().cloned().collect())
-            .unwrap_or_default()
-    }
-
-    fn schema(&self, name: &str) -> Option<Arc<dyn SchemaProvider>> {
-        self.schema_provider(name)
-    }
-
-    fn register_schema(
-        &self,
-        name: &str,
-        schema: Arc<dyn SchemaProvider>,
-    ) -> DFResult<Option<Arc<dyn SchemaProvider>>> {
-        self.register_schema_provider(name, schema)
-    }
-}
 
 #[async_trait]
 impl PartitionAwareCatalog for CayenneCatalogProvider {
@@ -288,7 +28,7 @@ impl PartitionAwareCatalog for CayenneCatalogProvider {
     ) -> crate::catalogconnector::Result<Option<String>> {
         let metadata_table_name = format!("{schema_name}/{table_name}");
         let metadata = self
-            .catalog
+            .metadata_catalog()
             .get_table(&metadata_table_name)
             .await
             .map_err(
@@ -300,408 +40,5 @@ impl PartitionAwareCatalog for CayenneCatalogProvider {
             )?;
 
         Ok(metadata.partition_column)
-    }
-}
-
-#[async_trait]
-impl RefreshableCatalogProvider for CayenneCatalogProvider {
-    async fn refresh(&self) -> std::result::Result<(), Box<dyn std::error::Error + Send + Sync>> {
-        let table_names = self.catalog.list_table_names().await.unwrap_or_else(|e| {
-            tracing::warn!("Failed to list existing Cayenne tables: {e}");
-            Vec::new()
-        });
-
-        // Group tables by namespace (tables stored as "namespace/table_name" in metadata).
-        let mut grouped: HashMap<String, Vec<String>> = HashMap::new();
-        for full_name in &table_names {
-            if let Some((ns, _table)) = full_name.split_once('/') {
-                grouped
-                    .entry(ns.to_string())
-                    .or_default()
-                    .push(full_name.clone());
-            } else {
-                tracing::debug!("Cayenne table '{full_name}' has no namespace prefix, skipping");
-            }
-        }
-
-        let existing_schemas = match self.schemas.read() {
-            Ok(schemas) => schemas.clone(),
-            Err(poisoned) => poisoned.into_inner().clone(),
-        };
-
-        let mut new_schemas: HashMap<String, Arc<dyn SchemaProvider>> = HashMap::new();
-        for (ns, full_names) in &grouped {
-            let refreshed_schema = CayenneSchemaProvider::try_new(
-                Arc::clone(&self.catalog),
-                ns,
-                full_names,
-                Arc::clone(&self.runtime_env),
-            )
-            .await?;
-
-            if let Some(existing_schema) = existing_schemas.get(ns)
-                && let Some(existing_cayenne_schema) = existing_schema
-                    .as_any()
-                    .downcast_ref::<CayenneSchemaProvider>()
-            {
-                existing_cayenne_schema
-                    .refresh_from(&refreshed_schema)
-                    .await;
-                new_schemas.insert(ns.clone(), Arc::clone(existing_schema));
-            } else {
-                new_schemas.insert(ns.clone(), Arc::new(refreshed_schema));
-            }
-        }
-
-        for (ns, existing_schema) in &existing_schemas {
-            if grouped.contains_key(ns) {
-                continue;
-            }
-
-            if let Some(existing_cayenne_schema) = existing_schema
-                .as_any()
-                .downcast_ref::<CayenneSchemaProvider>()
-            {
-                existing_cayenne_schema.clear_tables();
-            }
-        }
-
-        if !new_schemas.is_empty() {
-            let total_tables: usize = grouped.values().map(Vec::len).sum();
-            tracing::debug!(
-                "Loaded {total_tables} existing Cayenne table{} across {} namespace{}",
-                if total_tables == 1 { "" } else { "s" },
-                new_schemas.len(),
-                if new_schemas.len() == 1 { "" } else { "s" },
-            );
-        }
-
-        match self.schemas.write() {
-            Ok(mut schemas) => *schemas = new_schemas,
-            Err(poisoned) => *poisoned.into_inner() = new_schemas,
-        }
-
-        Ok(())
-    }
-}
-
-/// Schema provider for a single namespace within a Cayenne catalog.
-///
-/// Tables in the Cayenne metadata catalog are stored with namespace-prefixed
-/// names (`namespace/table_name`). This schema provider manages tables for
-/// one namespace, exposing them under their short (unqualified) names.
-pub struct CayenneSchemaProvider {
-    /// The underlying Cayenne metadata catalog.
-    catalog: Arc<dyn MetadataCatalog>,
-    /// The namespace this schema provider represents.
-    namespace: String,
-    /// Shared `RuntimeEnv` for cache coherence with the main runtime.
-    runtime_env: Arc<RuntimeEnv>,
-    /// Table providers keyed by short (unqualified) table name.
-    tables: RwLock<HashMap<String, Arc<dyn TableProvider>>>,
-}
-
-impl std::fmt::Debug for CayenneSchemaProvider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("CayenneSchemaProvider").finish()
-    }
-}
-
-impl CayenneSchemaProvider {
-    /// Create a new schema provider for a namespace, loading specified tables.
-    ///
-    /// `full_table_names` are namespace-prefixed names (`namespace/table_name`) as
-    /// stored in the Cayenne metadata catalog.
-    pub async fn try_new(
-        catalog: Arc<dyn MetadataCatalog>,
-        namespace: &str,
-        full_table_names: &[String],
-        runtime_env: Arc<RuntimeEnv>,
-    ) -> Result<Self> {
-        let ns_prefix = format!("{namespace}/");
-        let mut tables: HashMap<String, Arc<dyn TableProvider>> = HashMap::new();
-        for full_name in full_table_names {
-            let short_name = full_name.strip_prefix(&ns_prefix).unwrap_or(full_name);
-            match Self::load_table(&catalog, full_name, &runtime_env).await {
-                Ok(Some(provider)) => {
-                    tables.insert(short_name.to_string(), provider);
-                }
-                Ok(None) => {
-                    tracing::debug!(
-                        "Table '{full_name}' listed in catalog but could not be loaded"
-                    );
-                }
-                Err(e) => {
-                    tracing::warn!("Failed to load Cayenne table '{full_name}': {e}");
-                }
-            }
-        }
-
-        Ok(Self {
-            catalog,
-            namespace: namespace.to_string(),
-            runtime_env,
-            tables: RwLock::new(tables),
-        })
-    }
-
-    /// Create an empty schema provider for a namespace (used by DDL).
-    #[must_use]
-    pub fn new_empty(
-        catalog: Arc<dyn MetadataCatalog>,
-        namespace: String,
-        runtime_env: Arc<RuntimeEnv>,
-    ) -> Self {
-        Self {
-            catalog,
-            namespace,
-            runtime_env,
-            tables: RwLock::new(HashMap::new()),
-        }
-    }
-
-    fn tables_snapshot(&self) -> HashMap<String, Arc<dyn TableProvider>> {
-        match self.tables.read() {
-            Ok(tables) => tables.clone(),
-            Err(poisoned) => poisoned.into_inner().clone(),
-        }
-    }
-
-    fn replace_tables(&self, tables: HashMap<String, Arc<dyn TableProvider>>) {
-        match self.tables.write() {
-            Ok(mut existing_tables) => *existing_tables = tables,
-            Err(poisoned) => *poisoned.into_inner() = tables,
-        }
-    }
-
-    async fn refresh_from(&self, source: &Self) {
-        let existing_tables = self.tables_snapshot();
-        let refreshed_tables = source.tables_snapshot();
-        let mut merged_tables = HashMap::with_capacity(refreshed_tables.len());
-
-        for (table_name, refreshed_provider) in refreshed_tables {
-            let provider_to_use = if let Some(existing_provider) = existing_tables.get(&table_name)
-            {
-                if Self::refresh_table_provider_in_place(existing_provider, &refreshed_provider)
-                    .await
-                {
-                    Arc::clone(existing_provider)
-                } else {
-                    refreshed_provider
-                }
-            } else {
-                refreshed_provider
-            };
-
-            merged_tables.insert(table_name, provider_to_use);
-        }
-
-        self.replace_tables(merged_tables);
-    }
-
-    async fn refresh_table_provider_in_place(
-        existing_provider: &Arc<dyn TableProvider>,
-        refreshed_provider: &Arc<dyn TableProvider>,
-    ) -> bool {
-        // Case 1: Non-partitioned CayenneTableProvider — refresh in place from the
-        // freshly-loaded provider.
-        if let Some(existing_cayenne) = Self::cayenne_table_from_provider(existing_provider) {
-            let Some(refreshed_cayenne) = Self::cayenne_table_from_provider(refreshed_provider)
-            else {
-                return false;
-            };
-
-            if let Err(err) = existing_cayenne.refresh(refreshed_cayenne).await {
-                tracing::warn!("Failed to refresh Cayenne table provider in place: {err}");
-                return false;
-            }
-
-            return true;
-        }
-
-        // Case 2: PolyTableProvider — extract the writer and check its inner type.
-        let Some(poly) = existing_provider
-            .as_any()
-            .downcast_ref::<PolyTableProvider>()
-        else {
-            return false;
-        };
-
-        let existing_writer = poly.writer();
-
-        // Case 2a: Non-partitioned CayenneTableProvider inside PolyTableProvider.
-        if let Some(existing_cayenne) = existing_writer
-            .as_any()
-            .downcast_ref::<CayenneTableProvider>()
-        {
-            let refreshed_writer = refreshed_provider
-                .as_any()
-                .downcast_ref::<PolyTableProvider>()
-                .map(PolyTableProvider::writer);
-
-            let refreshed_cayenne = refreshed_writer
-                .as_ref()
-                .and_then(|w| w.as_any().downcast_ref::<CayenneTableProvider>());
-
-            if let Some(refreshed) = refreshed_cayenne {
-                if let Err(err) = existing_cayenne.refresh(refreshed).await {
-                    tracing::warn!("Failed to refresh Cayenne table provider in place: {err}");
-                    return false;
-                }
-                return true;
-            }
-
-            return false;
-        }
-
-        // Case 2b: PartitionTableProvider inside PolyTableProvider — refresh each
-        // partition's inner CayenneTableProvider from itself (reloads deletion
-        // vectors, protected snapshots, snapshot ID, and listing table from the
-        // catalog).
-        if let Some(partition_provider) = existing_writer
-            .as_any()
-            .downcast_ref::<PartitionTableProvider>()
-        {
-            let providers = partition_provider.partition_table_providers().await;
-            for provider in &providers {
-                // Each partition's inner provider is a CayenneTableProvider.
-                let Some(cayenne) = provider
-                    .as_any()
-                    .downcast_ref::<CayenneTableProvider>()
-                else {
-                    tracing::warn!(
-                        "Partition sub-provider is not a CayenneTableProvider, skipping refresh"
-                    );
-                    continue;
-                };
-                if let Err(err) = cayenne.refresh(cayenne).await {
-                    tracing::warn!("Failed to refresh partitioned Cayenne table in place: {err}");
-                    // Fail safely: keep the existing partitioned provider rather than
-                    // signaling failure that would cause it to be replaced by a
-                    // non-partitioned provider.
-                    return true;
-                }
-            }
-            return true;
-        }
-
-        false
-    }
-
-    fn cayenne_table_from_provider(
-        provider: &Arc<dyn TableProvider>,
-    ) -> Option<&CayenneTableProvider> {
-        provider.as_any().downcast_ref::<CayenneTableProvider>()
-    }
-
-    fn clear_tables(&self) {
-        self.replace_tables(HashMap::new());
-    }
-
-    /// Returns a reference to the underlying Cayenne metadata catalog.
-    #[must_use]
-    pub fn metadata_catalog(&self) -> &Arc<dyn MetadataCatalog> {
-        &self.catalog
-    }
-
-    /// Returns the namespace this schema provider represents.
-    #[must_use]
-    pub fn namespace(&self) -> &str {
-        &self.namespace
-    }
-
-    /// Construct the full metadata table name for a short table name.
-    fn full_table_name(&self, short_name: &str) -> String {
-        format!("{}/{short_name}", self.namespace)
-    }
-
-    /// Create a [`CayenneTableProvider`] for a table by its full metadata name.
-    async fn load_table(
-        catalog: &Arc<dyn MetadataCatalog>,
-        table_name: &str,
-        runtime_env: &Arc<RuntimeEnv>,
-    ) -> Result<Option<Arc<dyn TableProvider>>> {
-        // Check if the table exists in the catalog
-        match catalog.get_table(table_name).await {
-            Ok(_metadata) => {
-                let builder =
-                    CayenneTableProviderBuilder::new(Arc::clone(catalog), Arc::clone(runtime_env));
-
-                match builder.open(table_name).await {
-                    Ok(provider) => Ok(Some(Arc::new(provider) as Arc<dyn TableProvider>)),
-                    Err(e) => {
-                        tracing::warn!("Failed to open Cayenne table '{table_name}': {e}");
-                        Ok(None)
-                    }
-                }
-            }
-            Err(cayenne::catalog::CatalogError::TableNotFound { .. }) => Ok(None),
-            Err(e) => Err(Error::TableProvider { source: e }),
-        }
-    }
-}
-
-#[async_trait]
-impl SchemaProvider for CayenneSchemaProvider {
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn table_names(&self) -> Vec<String> {
-        self.tables
-            .read()
-            .map(|tables| tables.keys().cloned().collect())
-            .unwrap_or_default()
-    }
-
-    fn table_exist(&self, name: &str) -> bool {
-        self.tables
-            .read()
-            .map(|tables| tables.contains_key(name))
-            .unwrap_or(false)
-    }
-
-    async fn table(&self, name: &str) -> DFResult<Option<Arc<dyn TableProvider>>> {
-        // Check in-memory cache first
-        if let Ok(tables) = self.tables.read()
-            && let Some(provider) = tables.get(name)
-        {
-            return Ok(Some(Arc::clone(provider)));
-        }
-
-        // Try to load from catalog (lazy loading) using namespace-prefixed name
-        let full_name = self.full_table_name(name);
-        match Self::load_table(&self.catalog, &full_name, &self.runtime_env).await {
-            Ok(Some(provider)) => {
-                if let Ok(mut tables) = self.tables.write() {
-                    tables.insert(name.to_string(), Arc::clone(&provider));
-                }
-                Ok(Some(provider))
-            }
-            Ok(None) => Ok(None),
-            Err(e) => Err(datafusion::error::DataFusionError::External(Box::new(e))),
-        }
-    }
-
-    fn register_table(
-        &self,
-        name: String,
-        table: Arc<dyn TableProvider>,
-    ) -> DFResult<Option<Arc<dyn TableProvider>>> {
-        match self.tables.write() {
-            Ok(mut tables) => Ok(tables.insert(name, table)),
-            Err(_) => Err(datafusion::error::DataFusionError::Internal(
-                "Failed to acquire write lock on Cayenne tables".to_string(),
-            )),
-        }
-    }
-
-    fn deregister_table(&self, name: &str) -> DFResult<Option<Arc<dyn TableProvider>>> {
-        match self.tables.write() {
-            Ok(mut tables) => Ok(tables.remove(name)),
-            Err(_) => Err(datafusion::error::DataFusionError::Internal(
-                "Failed to acquire write lock on Cayenne tables".to_string(),
-            )),
-        }
     }
 }
