+++ crates/runtime/src/datafusion/cayenne_ddl/mod.rs
@@ -27,9 +27,10 @@ pub mod planner;
 
 use datafusion::catalog::CatalogProvider;
 
+use cayenne::CayenneCatalogProvider;
+
 use super::composed_catalog::ComposedCatalogProvider;
 use crate::catalogconnector::PartitionAwareCatalog;
-use crate::catalogconnector::cayenne::provider::CayenneCatalogProvider;
 
 /// Check whether the given catalog provider is a Cayenne-backed catalog.
 pub fn is_cayenne_catalog(provider: &dyn CatalogProvider) -> bool {
